+++ rsc/csharp/ParserExamples/BaseClass.cs
@@ -1,0 +1,15 @@
+
+namespace ParserExamples {
+
+    /// <summary>
+    /// Base class for this project.
+    /// </summary>
+    /// <threadsafety>
+    /// This class is mutable. And it is not thread-safe.
+    /// </threadsafety>
+    public abstract class BaseClass {
+
+        /// <summary>A string representation of this instance.</value>
+        public virtual string ToString();
+    }
+}