+++ rsc/java/ParserExamples/BaseClass.java
@@ -1,0 +1,11 @@
+package ParserExamples;
+
+/// <summary>
+/// Base class for this project.
+/// </summary>
+public abstract class AlbumInfo {
+
+	/// <summary>A string representation of this instance.</value>
+	public String toString();
+
+}